included new adapter plate to youbot-2
origin of new adapter plate, the centre of the mounting location, is 0.159 + 0.08
0.159 from the orignal centre of youbot rear place, which i removed, plus 8cm forward offset.

increased arm offset in z and y. the origin of arm_0 does not line up with the origin of the arm mount.
This offset to my best estimate +/- 2mm. I will add images to issue mas-group/mas_industrial_robotics#64

this should close #64