Cleanup, simply and syncronize all of the robot.urdf.xacro
All of the robots are now considered to be configured as they were
for German Open 2015. Which assumes we would have enough laser scanners,
rear platforms and grippers to go around. But the code-base expects frames,
sensors, etc. to exist so they would be modified to the same state before being run, so
this is a reasonable assumption.