[mir_actions][pick] Added ability to re-perceive and using different gripper open based on width of object
Re-perceive was added because the pose estimates for object at the edge
of the workstation were not accurate enough to pick small objects
successfully. With re-perceive enabled, the robot goes in front of the
object and perceives the workstation again. If it finds the object again it is
looking for, then it will use the pose estimate of this new perceive.
However, if it does not find the object in re-perceive, then it will
pick the object based on the old pose estimate (from the previous
perceive location action)